Add logic for Dynamic Slow container slices
