+++ facia/app/slices/Slice.scala
@@ -54,6 +54,30 @@ case object Ghidorah extends Slice
  */
 case object Anguirus extends Slice
 
+/* ._________________._________________.
+ * |_________________|_________________|
+ * |_________________|_________________|
+ * |_________________|_________________|
+ * |_________________|_________________|
+ */
+case object Minilla extends Slice
+
+/* ._________________.________.________.
+ * |_________________|########|########|
+ * |_________________|########|########|
+ * |_________________|        |        |
+ * |_________________|________|________|
+ */
+case object Destoroyah extends Slice
+
+/* ._________________._________________.
+ * |_________________|#################|
+ * |_________________|#################|
+ * |_________________|#################|
+ * |_________________|_________________|
+ */
+case object Gigan extends Slice
+
 /* .__________________________.________.
  * |        ##################|########|
  * |        ##################|########|
