+++ facia/test/slices/DynamicSlowTest.scala
@@ -1,0 +1,34 @@
+package slices
+
+import ArbitraryStories._
+import org.scalatest.OptionValues._
+import org.scalacheck.Arbitrary.arbitrary
+import common.Seqs._
+
+class DynamicSlowTest extends DynamicContainerTest {
+  override val slicesFor: (Seq[Story]) => Option[Seq[Slice]] = DynamicSlow.slicesFor
+
+  it should "for only standard items return Minilla" in {
+    forAll(storySeqGen(0)) { stories: Seq[Story] =>
+      whenever(stories.nonEmpty) {
+        slicesFor(stories).value.headOption.value shouldEqual Minilla
+      }
+    }
+  }
+
+  it should "for a single big and n standards, return Gigan" in {
+    forAll(storySeqGen(0), arbitrary[Boolean]) { (stories: Seq[Story], isBoosted: Boolean) =>
+      slicesFor(Story(1, isBoosted) +: stories).value.headOption.value shouldEqual Gigan
+    }
+  }
+
+  it should "for n > 1 bigs and m standards, return Destoroyah" in {
+    forAll(storySeqGen(1)) { stories: Seq[Story] =>
+      val bigs = stories.countWhile(_.group == 1)
+
+      whenever(bigs > 1) {
+        slicesFor(stories).value.headOption.value shouldEqual Destoroyah
+      }
+    }
+  }
+}
